backend/requirements.txt
fastapi
uvicorn
pandas
xgboost
scikit-learn
joblib
shap
google-generativeai
pydantic